requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
psycopg2-binary
python-dotenv
bcrypt==4.0.1
python-jose[cryptography]==3.3.0